requirements.txt
scikit-learn

pandas

sqlalchemy

requests

xgboost

tensorflow
matplotlib

catboost

beautifulsoup4

MLB-StatsApi

redis

shap